rename src directory and references
