+++ src/applications/edu-benefits/gi-bill-school-feedback/reducers/schoolSelect.js
