+++ src/applications/edu-benefits/gi-bill-school-feedback/selectors/schoolSearch.js
