+++ src/applications/edu-benefits/tests/complaint-tool/00.complaint-tool.e2e.spec.js
@@ -1,10 +1,10 @@
 const E2eHelpers = require('../../../../platform/testing/e2e/helpers');
-const manifest = require('../../complaint-tool/manifest.json');
+const manifest = require('../../gi-bill-school-feedback/manifest.json');
 const Timeouts = require('../../../../platform/testing/e2e/timeouts.js');
 
 module.exports = E2eHelpers.createE2eTest((client) => {
   client
-    .url(`${E2eHelpers.baseUrl}/education/complaint-tool/form`)
+    .url(`${E2eHelpers.baseUrl}/education/gi-bill-school-feedback/form`)
     .waitForElementVisible('body', Timeouts.normal)
     .axeCheck('.main');
