+++ src/applications/edu-benefits/tests/complaint-tool/config/benefitsInfo.unit.spec.js
@@ -4,9 +4,9 @@ import sinon from 'sinon';
 import { mount } from 'enzyme';
 
 import { DefinitionTester, selectCheckbox } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../complaint-tool/config/form';
+import formConfig from '../../../gi-bill-school-feedback/config/form';
 
-describe('complaint tool benefits info', () => {
+describe('gi bill school feedback benefits info', () => {
   const { schema, uiSchema } = formConfig.chapters.benefitsInformation.pages.benefitsInformation;
 
   it('should render', () => {
