+++ src/applications/edu-benefits/tests/complaint-tool/config/schoolInfo.unit.spec.js
@@ -7,9 +7,9 @@ import {
   DefinitionTester,
   fillData
 } from '../../../../../platform/testing/unit/schemaform-utils.jsx';
-import formConfig from '../../../complaint-tool/config/form';
+import formConfig from '../../../gi-bill-school-feedback/config/form';
 
-describe('complaint tool school info', () => {
+describe('gi bill school feedback school info', () => {
   const { schema, uiSchema } = formConfig.chapters.schoolInformation.pages.schoolInformation;
   _.unset(uiSchema, 'educationDetails.school.facilityCode');
