+++ src/applications/personalization/dashboard/helpers.jsx
@@ -14,7 +14,7 @@ export const formBenefits = {
   '22-5495': 'education benefits',
   '40-10007': 'pre-need determination of eligibility in a VA national cemetery',
   VIC: 'Veteran ID Card',
-  'complaint-tool': 'feedback',
+  'gi-bill-school-feedback': 'feedback',
   '21-686C': 'dependent status'
 };
 
@@ -46,7 +46,7 @@ export const formLinks = {
   '22-5495': '/education/apply-for-education-benefits/application/5495/',
   '40-10007': '/burials-and-memorials/pre-need/form-10007-apply-for-eligibility/',
   VIC: '/veteran-id-card/apply/',
-  'complaint-tool': '/education/complaint-tool/form',
+  'gi-bill-school-feedback': '/education/gi-bill-school-feedback/form',
   '21-686C': '/disability-benefits/686/dependent-status/'
 };
 
@@ -82,7 +82,7 @@ export const sipEnabledForms = new Set([
   '22-5495',
   '40-10007',
   'VIC',
-  'complaint-tool'
+  'gi-bill-school-feedback'
 ]);
